backend/requirements.txt
flask
openai-whisper
sqlalchemy